Added casting to the GetRawValue method so the correct type can be realized.

+++ src/BoboBrowse.Net/Query/RecencyBoostScorerBuilder.cs
@@ -115,7 +115,8 @@ namespace BoboBrowse.Net.Query
             public override float Score()
             {
                 float rawScore = _innerScorer.Score();
-                long timeVal = _termList.GetRawValue(_orderArray.Get(_innerScorer.DocID()));
+                // TODO: Try to find a way to get this value without having to cast.
+                long timeVal = (long)_termList.GetRawValue(_orderArray.Get(_innerScorer.DocID()));
                 float timeScore = _parent.ComputeTimeFactor(timeVal);
                 return RecencyBoostScorerBuilder.CombineScores(timeScore, rawScore);
             }
